IBP-1420 - blink_permission_id INT to BIGINT

+++ sql/Pg-upgrade-2.8.42-2.8.43.sql
@@ -1,4 +1,4 @@
 ALTER TABLE bank_account
-    ADD COLUMN blink_permission_id INT;
+    ADD COLUMN blink_permission_id BIGINT;
 
 UPDATE defaults SET fldvalue = '2.8.43' WHERE fldname = 'version';
